StealthsystemV1
Hello everyone,
These are the stealth files. So far, you can hide in a crowd (see GIF), where I found that only from 2 people is to be regarded as a crowd. If you crouch down the search area of ​​the AI ​​and you are therefore harder to see. What else did you have to add to this?